Add test for FindShelter

+++ nontserver/test/LoginAndSignup/NontOwnerUpdate.test.js
@@ -145,9 +145,16 @@ describe("Nont Owner Update", () => {
 
 describe("Clear Up", () => {
   it("Clear up", (done) => {
-    NontOwner.deleteOne({
-      email: "nontOwnerTestUpdate@kojira.com",
-    }).then((err) => done());
+    NontOwner.deleteOne({ email: "nontOwnerTestUpdate@kojira.com" }).then(
+      (result) => {
+        NontOwner.findOne({ _id: id }).then(
+          (result2) => {
+            expect(result2).to.be.null;
+            done()
+          }
+        )
+      }
+    );
   });
 });
